Fix fully played logic

+++ src/main/java/net/pms/dlna/DLNAResource.java
@@ -3117,7 +3117,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 	/**
 	 * @return The system time when the resource was last (re)started
 	 */
-	public double getLastStartSystemTime() {
+	public long getLastStartSystemTime() {
 		return lastStartSystemTime;
 	}
 
@@ -3148,7 +3148,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 	 * @return The system time when the resource was last (re)started
 	 *         by a user.
 	 */
-	public double getLastStartSystemTimeUser() {
+	public long getLastStartSystemTimeUser() {
 		return lastStartSystemTimeUser;
 	}
